tweaked animations, changed input settings
I swapped the character over to a walk animation so that it better matched up with the speed it was going. I increased the deadzone of the right joystick so that it didn't flip the character when reverting to a neutral position. Note that the right joystick is inverted on the Y axis